HoPeRaTiOn EaLtH
Okey this might seem like  lot of stuff, and actually it's a lot of changed stuff. I included a lot of fixes, AND deleted many many unused files. This severely decreases mission filesize. there is still EOS and EPD which is not used I think, delete that too? This version is not yet tested thoroughly so please don't upload it to master branch.  But you are free to comment if you have questions.